Resolve "Updated implementation of technical questions total mark in interview form"

See merge request edgsolutions-engineering/clear-fusion!555

+++ HumanitarianAssistance.Application/Project/Commands/Create/AddInterviewDetailsCommandHandler.cs
@@ -29,7 +29,7 @@ namespace HumanitarianAssistance.Application.Project.Commands.Create {
                 Description = request.Description,
                 NoticePeriod = request.NoticePeriod,
                 AvailableDate = request.AvailableDate,
-                WrittenTestMarks = (int) request.WrittenTestMarks,
+                WrittenTestMarks = request.WrittenTestMarks,
                 CurrentBase = request.CurrentBase,
                 CurrentOther = request.CurrentOther,
                 ExpectationBase = request.ExpectationBase,
